backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pymongo==4.6.0
python-socketio==5.10.0
python-dotenv==1.0.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
